refactor: move lifecycle hooks to tool modules and simplify config format
