+++ tests/integration/test_support/tools.py
@@ -1,5 +1,13 @@
 from typing import Dict, Any, Optional
 from google.adk.tools import ToolContext
+from pathlib import Path
+from solace_ai_connector.common.log import log
+from tests.integration.test_support.lifecycle_tracker import track
+
+if "SamAgentComponent" not in globals():
+    from solace_agent_mesh.agent.sac.component import SamAgentComponent
+if "AnyToolConfig" not in globals():
+    from solace_agent_mesh.agent.tools.tool_config_types import AnyToolConfig
 
 
 async def get_weather_tool(
@@ -22,3 +30,27 @@ async def get_weather_tool(
             "unit": unit or "celsius",
             "condition": "unknown",
         }
+
+
+async def yaml_init_hook(component: "SamAgentComponent", tool_config: "AnyToolConfig"):
+    """A simple init hook for YAML configuration tests."""
+    log.info("yaml_init_hook called.")
+    tracker_file = Path(tool_config.tool_config["tracker_file"])
+    track(tracker_file, "yaml_init_called")
+
+
+async def yaml_cleanup_hook(
+    component: "SamAgentComponent", tool_config: "AnyToolConfig"
+):
+    """A simple cleanup hook for YAML configuration tests."""
+    log.info("yaml_cleanup_hook called.")
+    tracker_file = Path(tool_config.tool_config["tracker_file"])
+    track(tracker_file, "yaml_cleanup_called")
+
+
+async def failing_init_hook(
+    component: "SamAgentComponent", tool_config: "AnyToolConfig"
+):
+    """An init hook that always fails."""
+    log.info("failing_init_hook called, will raise ValueError.")
+    raise ValueError("Simulated fatal init failure")
